+++ test/crudTests.test.js
@@ -6,44 +6,9 @@ const request = supertest(app);
 const axios = require('axios');
 import 'babel-polyfill';
 
-const fakeData = [
-  {
-    itemId: '200',
-    title: 'Dog Toy',
-    description:
-      'Lorem ipsum dolor sit amet. Consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
-    SKU: '2342048',
-    primaryBrand: 'PetCo',
-    daysToShip: 'Ships In Two Business Days',
-    directions:
-      'CAUTION: Intended for pet use only. Not a childs toy. Choose toys based on dogs playing habits. Toys should be large enough to not be swallowed. IMPORTANT: No pet toy is indestructible. Small parts present a choking or gastrointestinal blockage risk. Always supervise your pet during play to prevent accidental swallowing of parts. Inspsect toy regularly and replace if any part becomes loose. If toy becomes wet, some color transfer may occur. Spot clean only.',
-    primaryColor: 'Multicolor',
-    material: 'Plush',
-    length: '8 IN',
-    width: '2 IN',
-    additionalDetails:
-      'Lorem ipsum dolor sit amet. Consectetur adipiscing elit.',
-  },
-];
-
-beforeEach(async () => {
-  await db.Description.create(fakeData);
-});
-
-afterEach(async () => {
-  await db.Description.deleteOne({ itemId: 200 });
-});
-
-describe('CRUD Endpoint Tests', () => {
-  it('deletes an object at the /descriptionObject delete endpoint', async (done) => {
-    const oldCount = await db.Description.count({});
-    const response = await request.delete(
-      `/descriptionObject/${oldCount + 100}`
-    );
-    const newCount = await db.Description.count({});
-    expect(response.status).toBe(200);
-    expect(newCount).toEqual(oldCount - 1);
-    done();
+describe('CRUD Create Endpoint Tests', () => {
+  beforeEach(async () => {
+    await db.Description.deleteMany({ itemId: 200 });
   });
 
   it('posts to the /descriptionObject endpoint', async (done) => {
@@ -52,18 +17,7 @@ describe('CRUD Endpoint Tests', () => {
     const newCount = await db.Description.count({});
     expect(response.status).toBe(200);
     expect(newCount).toEqual(oldCount + 1);
-    await request.delete(`/descriptionObject/${newCount + 100}`);
-    done();
-  });
-
-  it('modifies an object at the /descriptionObject/:itemId put endpoint', async (done) => {
-    const itemSearchingFor = await db.Description.find({ itemId: 200 });
-    const oldTitle = itemSearchingFor[0].title;
-    await request.put('/descriptionObject/200').send({ title: 'Updated!' });
-    const sameItemSearch = await db.Description.find({ itemId: 200 });
-    const newTitle = sameItemSearch[0].title;
-    expect(oldTitle).toBe('Dog Toy');
-    expect(newTitle).toBe('Updated!');
+    await request.delete(`/descriptionObject/${oldCount + 100}`);
     done();
   });
 });
